Implementation of command FIND and TYPE
* Implemeted new functions within libDos9 in order to sort matching file and
non-matching files more easily (typically, in a single call).

* Reimplemented the TYPE command in order to get something compatible with
cmd.exe (for the *very* majority of scipts).

* Implemented the FIND command as a internal command. It matches closely
cmd.exe behavior, execpt for some inconsistencies I've tried to address.

* Implemented the %* variable. That lacked for long time in the Dos9
implementation. There is still little concerns, like the facet that under
Unix-like operating systems, it is not possible to get the original line,
thus, Dos9 tries to reconstruct it the best it can, but it is not optimal