Added locks to pBat_RunFile() functions, fixed inheritance issues
Added thread locking mechanisms for the pBat_RunFile() functions in order
to provide safer thread handling of file descriptors on windows. The
support is only partial for now but but PBAT_RUNFILE_LOCK() and
PBAT_RUNFILE_RELEASE() should be used when openning files to prevent
inheritance issues

Removed most references to pBat_SetFdInheritance() because of changes in
handling fork() and CreateProcess() calls. On windows, all the handles
execpt the ones needed are set to be not inheritable just before calls
to CreateProcess(). This creates a minimal speed penalty. On *nix,
instead of reseting inheritance using FD_CLOEXEC, every unnecessary file
is closed right after the fork() call.